Extract OutputTitle to its own file
